WIP Replace .zip with a self-extracting 7-zip (#87)
WIP NOTE: README.md does yet reflect the change in distribution format.
Perhaps the SFX icon should combine the 7-zip icon with the w64devkit
icon. It's currently just the w64devkit icon.

Cuts the distribution size by around 60%, and unpacks around 100x faster
than Windows built-in .zip support. However, 7-zip is the worst packaged
project included so far. The source release is a tarbomb with a URL that
contains two variations of the release version. It is not designed for
cross-compilation, and so requires minor source corrections and a custom
build script. Still, considering the circumstances, it's surprising how
easily it compiles.